+++ Ethereum/controllers/ethereum.js
@@ -554,8 +554,9 @@ exports.getSearch = async (req, res) => {
 
 exports.postSendSignedTransaction = async function (req, res) {
   var raw = req.body.raw;
-
   try {
+    if (String(raw).substring(0, 2) != '0x') raw = `0x${raw}`;
+
     await web3.eth.sendSignedTransaction(raw)
       .on('receipt', (receipt) => {
         // console.log(JSON.stringify(receipt));
@@ -563,6 +564,7 @@ exports.postSendSignedTransaction = async function (req, res) {
           return res.json({ status: 200, msg: "success", data: receipt });
         return res.json({ status: 400, msg: "Empty receipt !" });
       });
+    return res.json({ status: 400, msg: "send error !" });
   } catch (error) {
     return res.json({ status: 400, msg: "Error !", data: error });
   }
