+++ Neo/config/index.js
@@ -16,9 +16,9 @@ module.exports = {
     db: "mongodb://localhost:27017/neo-db",
 
     // CRON_SLEEP_TIME: 500,	//if we decrease this to 100ms, cronservice has trouble after some time.
-    TX_CRON_TIME: 8 * 1000,	//if we decrease this to 100ms, cronservice has trouble after some time.
+    TX_CRON_TIME: 1000,	//if we decrease this to 100ms, cronservice has trouble after some time.
     TX_CRON_BLOCK_COUNT: 100, // block count per one running
 
-    ADDR_CRON_TIME: 10 * 1000,	//if we decrease this to 100ms, cronservice has trouble after some time.
+    ADDR_CRON_TIME: 1000,	//if we decrease this to 100ms, cronservice has trouble after some time.
     ADDR_CRON_TX_COUNT: 100 // tx count per one running
 }
